+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ml_inference/text_expansion_callout.test.tsx
@@ -8,18 +8,13 @@
 import { setMockValues } from '../../../../../__mocks__/kea_logic';
 import React from 'react';
 import { shallow } from 'enzyme';
-import { EuiText } from '@elastic/eui';
 import { HttpError } from '../../../../../../../common/types/api';
-import {
-  TextExpansionCallOut,
-  TextExpansionDismissButton,
-  ModelStarted,
-  FineTuneModelsButton,
-} from './text_expansion_callout';
+import { TextExpansionCallOut } from './text_expansion_callout';
 import { DeployModel } from './deploy_model';
 import { TextExpansionErrors } from './text_expansion_errors';
 import { ModelDeploymentInProgress } from './model_deployment_in_progress';
 import { ModelDeployed } from './model_deployed';
+import { ModelStarted } from './model_started';
 
 jest.mock('./text_expansion_callout_data', () => ({
   useTextExpansionCallOutData: jest.fn(() => ({
@@ -90,32 +85,5 @@ describe('TextExpansionCallOut', () => {
 
     const wrapper = shallow(<TextExpansionCallOut />);
     expect(wrapper.find(ModelStarted).length).toBe(1);
-  });
-
-  describe('ModelStarted', () => {
-    it('renders dismiss button if it is set to dismissable', () => {
-      const wrapper = shallow(
-        <ModelStarted dismiss={() => {}} isCompact={false} isDismissable isSingleThreaded />
-      );
-      expect(wrapper.find(TextExpansionDismissButton).length).toBe(1);
-    });
-    it('does not render dismiss button if it is set to non-dismissable', () => {
-      const wrapper = shallow(
-        <ModelStarted dismiss={() => {}} isCompact={false} isDismissable={false} isSingleThreaded />
-      );
-      expect(wrapper.find(TextExpansionDismissButton).length).toBe(0);
-    });
-    it('renders fine-tune button if the model is running single-threaded', () => {
-      const wrapper = shallow(
-        <ModelStarted dismiss={() => {}} isCompact={false} isDismissable isSingleThreaded />
-      );
-      expect(wrapper.find(FineTuneModelsButton).length).toBe(1);
-    });
-    it('does not render description if it is set to compact', () => {
-      const wrapper = shallow(
-        <ModelStarted dismiss={() => {}} isCompact isDismissable isSingleThreaded />
-      );
-      expect(wrapper.find(EuiText).length).toBe(1); // Title only
-    });
   });
 });
