+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/pipelines/ml_inference/text_expansion_callout.tsx
@@ -7,15 +7,7 @@
 
 import React from 'react';
 import { useValues } from 'kea';
-import {
-  EuiButtonEmpty,
-  EuiButtonIcon,
-  EuiCallOut,
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiIcon,
-  EuiText,
-} from '@elastic/eui';
+import { EuiButtonEmpty, EuiButtonIcon } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { KibanaLogic } from '../../../../../shared/kibana';
 import { IndexViewLogic } from '../../index_view_logic';
@@ -25,6 +17,7 @@ import { TextExpansionErrors } from './text_expansion_errors';
 import { DeployModel } from './deploy_model';
 import { ModelDeployed } from './model_deployed';
 import { ModelDeploymentInProgress } from './model_deployment_in_progress';
+import { ModelStarted } from './model_started';
 
 export interface TextExpansionCallOutState {
   dismiss: () => void;
@@ -76,114 +69,6 @@ export const FineTuneModelsButton: React.FC = () => (
       }
     )}
   </EuiButtonEmpty>
-);
-
-export const ModelStarted = ({
-  dismiss,
-  isCompact,
-  isDismissable,
-  isSingleThreaded,
-}: Pick<
-  TextExpansionCallOutState,
-  'dismiss' | 'isCompact' | 'isDismissable' | 'isSingleThreaded'
->) => (
-  <EuiCallOut color="success">
-    <EuiFlexGroup direction="column" gutterSize="s">
-      <EuiFlexItem grow>
-        <EuiFlexGroup direction="row" gutterSize="s" alignItems="center">
-          <EuiFlexItem grow={false}>
-            <EuiIcon type="checkInCircleFilled" color="success" />
-          </EuiFlexItem>
-          <EuiFlexItem grow>
-            <EuiText color="success" size="xs">
-              <h3>
-                {isSingleThreaded
-                  ? isCompact
-                    ? i18n.translate(
-                        'xpack.enterpriseSearch.content.index.pipelines.textExpansionCallOut.startedSingleThreadedTitleCompact',
-                        { defaultMessage: 'Your ELSER model is running single-threaded.' }
-                      )
-                    : i18n.translate(
-                        'xpack.enterpriseSearch.content.index.pipelines.textExpansionCallOut.startedSingleThreadedTitle',
-                        { defaultMessage: 'Your ELSER model has started single-threaded.' }
-                      )
-                  : isCompact
-                  ? i18n.translate(
-                      'xpack.enterpriseSearch.content.index.pipelines.textExpansionCallOut.startedTitleCompact',
-                      { defaultMessage: 'Your ELSER model is running.' }
-                    )
-                  : i18n.translate(
-                      'xpack.enterpriseSearch.content.index.pipelines.textExpansionCallOut.startedTitle',
-                      { defaultMessage: 'Your ELSER model has started.' }
-                    )}
-              </h3>
-            </EuiText>
-          </EuiFlexItem>
-          {isDismissable && (
-            <EuiFlexItem grow={false}>
-              <TextExpansionDismissButton dismiss={dismiss} />
-            </EuiFlexItem>
-          )}
-        </EuiFlexGroup>
-      </EuiFlexItem>
-      {!isCompact && (
-        <>
-          <EuiFlexItem grow>
-            <EuiText size="s">
-              <p>
-                {isSingleThreaded
-                  ? i18n.translate(
-                      'xpack.enterpriseSearch.content.index.pipelines.textExpansionCallOut.startedSingleThreadedBody',
-                      {
-                        defaultMessage:
-                          'This single-threaded configuration is great for testing your custom inference pipelines, however performance should be fine-tuned for production.',
-                      }
-                    )
-                  : i18n.translate(
-                      'xpack.enterpriseSearch.content.index.pipelines.textExpansionCallOut.startedBody',
-                      {
-                        defaultMessage:
-                          'Enjoy the power of ELSER in your custom Inference pipeline.',
-                      }
-                    )}
-              </p>
-            </EuiText>
-          </EuiFlexItem>
-          <EuiFlexItem>
-            <EuiFlexGroup
-              direction="row"
-              gutterSize="m"
-              alignItems="center"
-              justifyContent="flexStart"
-            >
-              <EuiFlexItem grow={false}>
-                {isSingleThreaded ? (
-                  <FineTuneModelsButton />
-                ) : (
-                  <EuiButtonEmpty
-                    iconSide="left"
-                    iconType="wrench"
-                    onClick={() =>
-                      KibanaLogic.values.navigateToUrl(TRAINED_MODELS_PATH, {
-                        shouldNotCreateHref: true,
-                      })
-                    }
-                  >
-                    {i18n.translate(
-                      'xpack.enterpriseSearch.content.indices.pipelines.textExpansionCallOut.viewModelsButton',
-                      {
-                        defaultMessage: 'View details',
-                      }
-                    )}
-                  </EuiButtonEmpty>
-                )}
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          </EuiFlexItem>
-        </>
-      )}
-    </EuiFlexGroup>
-  </EuiCallOut>
 );
 
 export const TextExpansionCallOut: React.FC<TextExpansionCallOutProps> = (props) => {
